Spanner: Changed the names of the PG sample files
