+++ spanner/src/pg_partitioned_dml.php
@@ -23,38 +23,31 @@
 
 namespace Google\Cloud\Samples\Spanner;
 
-// [START spanner_postgresql_dml_with_parameters]
+// [START spanner_postgresql_partitioned_dml]
 use Google\Cloud\Spanner\SpannerClient;
-use Google\Cloud\Spanner\Transaction;
 
 /**
- * Execute a batch of DML statements on a Spanner PostgreSQL database
+ * Execute Partitioned DML on a Spanner PostgreSQL database.
+ * See also https://cloud.google.com/spanner/docs/dml-partitioned.
  *
  * @param string $instanceId The Spanner instance ID.
  * @param string $databaseId The Spanner database ID.
  */
-function pg_spanner_dml_with_params(string $instanceId, string $databaseId): void
+function pg_partitioned_dml(string $instanceId, string $databaseId): void
 {
     $spanner = new SpannerClient();
     $instance = $spanner->instance($instanceId);
     $database = $instance->database($databaseId);
 
-    $database->runTransaction(function (Transaction $t) {
-        $count = $t->executeUpdate('INSERT INTO Singers (SingerId, "FirstName", "LastName")'
-            . ' VALUES ($1, $2, $3)',
-            [
-                'parameters' => [
-                    'p1' => 3,
-                    'p2' => 'Alice',
-                    'p3' => 'Henderson'
-                ]
-            ]);
-        $t->commit();
+    // Spanner PostgreSQL has the same transaction limits as normal Spanner. This includes a
+    // maximum of 20,000 mutations in a single read/write transaction. Large update operations can
+    // be executed using Partitioned DML. This is also supported on Spanner PostgreSQL.
+    // See https://cloud.google.com/spanner/docs/dml-partitioned for more information.
+    $count = $database->executePartitionedUpdate('DELETE FROM Singers WHERE "FirstName" IS NOT NULL');
 
-        printf('Inserted %s singer(s).' . PHP_EOL, $count);
-    });
+    printf('Deleted %s rows.' . PHP_EOL, $count);
 }
-// [END spanner_postgresql_dml_with_parameters]
+// [END spanner_postgresql_partitioned_dml]
 
 // The following 2 lines are only needed to run the samples
 require_once __DIR__ . '/../../testing/sample_helpers.php';
